Move convert_ad_dtype_to_np from nd_plug to utils
